+++ api/src/helpers/db.js
@@ -1,0 +1,17 @@
+const { executeSQL } = require('../mariadb')
+
+async function createCitizen(handle) {
+    console.log("Creating citizen: " + handle)
+    // try to load citizen from DB
+    const rows = await executeSQL("SELECT * FROM citizen WHERE handle=?", [handle])
+
+    if(rows.length === 0) {
+        // if no record, add new record
+        await executeSQL("INSERT INTO citizen (handle) values (?)", [handle])
+        await syncCitizen(handle)
+    }
+}
+
+module.exports = {
+    createCitizen
+}
