+++ api/src/modules/citizen/model.js
@@ -1,11 +1,8 @@
-const axios = require("axios")
-const cheerio = require('cheerio')
-const jwt = require('jsonwebtoken')
 
-const { manager } = require('../manager')
 const { executeSQL } = require('../mariadb')
 
 const { getUser } = require('../user/model')
+const { fetchCitizen } = require('../helpers/rsi')
 
 /*
 *   GET /citizen/<handle>
@@ -86,29 +83,7 @@ async function loadCitizenLocation(handle) {
     return home
 }
 
-async function fetchCitizen(handle) {
-    console.log('fetching citizen...')
-    try {
-        const baseURI = 'https://robertsspaceindustries.com'
-        const resp = await axios.get(baseURI + '/citizens/' + handle)
-        const $ = cheerio.load(resp.data)
-        info = {}
-        info.handle = handle
-        info.record = $('span:contains("UEE Citizen Record")', '#public-profile').next().text()
-        info.name = $('div.profile.left-col', '#public-profile').find('div.info').find('p.entry').find('strong.value').html()
-        info.bio = $('span:contains("Bio")', '#public-profile').next().text()
-        info.enlisted = $("span:contains('Enlisted')", '#public-profile').next().text()
-        info.portrait = baseURI + $('div.thumb', '#public-profile').children()[0].attribs.src
-        info.org = $('span:contains("Spectrum Identification (SID)")', '#public-profile').next().text()
-        info.orgRank = $('span:contains("Organization rank")', '#public-profile').next().text()
-        info.website = $('span:contains("Website")', '#public-profile').next().attr('href') || ''
-        info.verified = 0
-        return info
-    } catch (error) {
-        console.error(error)
-        return null
-    }
-}
+
 
 async function getInfo(handle) {
     citizen = await getCitizen(handle)
@@ -179,4 +154,4 @@ module.exports = {
     getLocation,
     setLocation,
     createCitizen
-};
+}
