+++ api/src/modules/user/model.js
@@ -1,9 +1,8 @@
-const { getCitizen, fetchCitizen, createCitizen } = require('../citizen/model')
+const { fetchCitizen } = require('../helpers/rsi')
+const { createCitizen } = require('../helpers/db')
 const { getVerificationCode, setVerified } = require('../verification')
 
 const{ manager } = require('../manager')
-
-
 
 
 async function getUser(usr) {
